Fix configuration system & allow building
