+++ build-logic/src/main/groovy/multiloader-common.gradle
@@ -42,7 +42,7 @@ processResources {
             'minecraft_version_range'      : minecraft_version_range,
             'fabric_version'               : fabric_version,
             'fabric_loader_version'        : fabric_loader_version,
-            'fabric_cloth_version'         : fabric_cloth_version,
+            'cloth_version'                : cloth_version,
             'fabric_modmenu_version'       : fabric_modmenu_version,
             'mod_name'                     : mod_name,
             'mod_author'                   : mod_author,
