+++ common/build.gradle
@@ -3,6 +3,10 @@ plugins {
     id 'net.neoforged.moddev'
 }
 
+repositories {
+    maven { url "https://maven.shedaniel.me/" }
+}
+
 neoForge {
     neoFormVersion = neo_form_version
     // Automatically enable AccessTransformers if the file exists
@@ -18,9 +22,14 @@ dependencies {
     // https://github.com/neoforged/NeoForge/blob/26.2.x/gradle.properties#L37
     // https://github.com/FabricMC/fabric-loader/blob/master/gradle.properties#L12
     compileOnly('net.fabricmc:sponge-mixin:0.17.3+mixin.0.8.7')
+
     // Fabric and NeoForge both bundle MixinExtras, so it is safe to use it in common
     compileOnly(annotationProcessor('io.github.llamalad7:mixinextras-common:0.5.3'))
 
+    compileOnly("me.shedaniel.cloth:cloth-config-fabric:${cloth_version}") {
+        exclude(group: "net.fabricmc", module: "fabric-loader")
+        exclude(group: "net.fabricmc.fabric-api")
+    }
 }
 
 configurations {
