+++ common/src/main/java/me/daxanius/npe/NoPryingEyesCommon.java
@@ -18,4 +18,8 @@ public class NoPryingEyesCommon {
 
         LOG.info(s);
     }
+
+    public static void initialize() {
+        NoPryingEyesConfig.load();
+    }
 }
