+++ fabric/src/main/java/me/daxanius/npe/ModMenuIntegration.java
@@ -1,0 +1,17 @@
+package me.daxanius.npe;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import me.daxanius.npe.config.ClothConfigScreenFactory;
+import net.fabricmc.loader.api.FabricLoader;
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        if (!FabricLoader.getInstance().isModLoaded("cloth-config")) {
+            return null;
+        }
+
+        return ClothConfigScreenFactory::build;
+    }
+}
