+++ fabric/src/main/java/me/daxanius/npe/NoPryingEyes.java
@@ -6,6 +6,6 @@ public class NoPryingEyes implements ModInitializer {
 
     @Override
     public void onInitialize() {
-
+        NoPryingEyesCommon.initialize();
     }
 }
