+++ gradle.properties
@@ -16,6 +16,8 @@ credits=TechPro242
 description=Provides full control over your data; namely telemetry and chat reporting.
 minecraft_version_range=[26.2, 26.3)
 
+cloth_version=26.2.155
+
 ## This is the version of minecraft that the 'common' project uses, you can find a list of all versions here
 ## https://projects.neoforged.net/neoforged/neoform
 neo_form_version=26.2-1
@@ -24,7 +26,6 @@ neo_form_version=26.2-1
 fabric_version=0.152.1+26.2
 fabric_loader_version=0.19.3
 
-fabric_cloth_version=26.2.155
 fabric_modmenu_version=20.0.0
 
 # NeoForge, see https://projects.neoforged.net/neoforged/neoforge for new versions
