+++ neoforge/src/main/java/me/daxanius/npe/NoPryingEyes.java
@@ -5,8 +5,7 @@ import net.neoforged.fml.common.Mod;
 
 @Mod(NoPryingEyesCommon.MOD_ID)
 public class NoPryingEyes {
-
     public NoPryingEyes(IEventBus eventBus) {
-
+        NoPryingEyesCommon.initialize();
     }
 }
