+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -8,7 +8,7 @@ version = "${version}" #mandatory
 displayName = "${mod_name}" #mandatory
 #updateJSONURL="https://change.me.example.invalid/updates.json" #optional (see https://docs.neoforged.net/docs/misc/updatechecker/)
 #displayURL="https://change.me.to.your.mods.homepage.example.invalid/" #optional (displayed in the mod UI)
-logoFile="${mod_id}.png"
+logoFile="assets/npe/icon.png"
 credits="${credits}"
 authors = "${mod_author}"
 description = '''${description}'''
@@ -24,6 +24,7 @@ type="required" #mandatory (Can be one of "required", "optional", "incompatible"
 versionRange = "[${neoforge_version},)" #mandatory
 ordering = "NONE" # The order that this dependency should load in relation to your mod, required to be either 'BEFORE' or 'AFTER' if the dependency is not mandatory
 side = "BOTH" # Side this dependency is applied on - 'BOTH', 'CLIENT' or 'SERVER'
+
 [[dependencies.${mod_id}]]
 modId = "minecraft"
 type="required" #mandatory (Can be one of "required", "optional", "incompatible" or "discouraged")
